# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /preview:
    post:
      operationId: savePreview
      tags: [preview]
      summary: Save HTML preview for a conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavePreviewInput"
      responses:
        "201":
          description: Saved preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreviewMeta"
  /preview/{id}:
    get:
      operationId: getPreview
      tags: [preview]
      summary: Serve the preview HTML page
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: HTML page
          content:
            text/html: {}
        "404":
          description: Not found
components:
  schemas:
...
    SavePreviewInput:
      type: object
      properties:
        conversationId:
          type: integer
        htmlContent:
          type: string
      required:
        - conversationId
        - htmlContent
    PreviewMeta:
      type: object
      properties:
        id:
          type: integer
        previewUrl:
          type: string
      required:
        - id
        - previewUrl